Replace $wpdb->prepare with an array_map to absint.
It's not possible to use prepare with an imploded array of numbers because prepare adds quotes around strings, which breaks the SQL WHERE IN check against the ancestor posts. Since we can't use prepare to avoid SQL injection, we are using absint on all the elements of the array before imploding it to make sure the only possible thing we could be including in our $wpdb->get_results call is integers.